Added .gitignore and renamed the package to com.github.trojanrobotics
