+++ resources/META-INF/MANIFEST.MF
@@ -1,6 +1,6 @@
 MIDlet-Name: ServoTesting
 MIDlet-Version: 1.0.0
 MIDlet-Vendor: FIRST
-MIDlet-1: RobotTemplate, , edu.wpi.first.wpilibj.templates.RobotTemplate
+MIDlet-1: RobotTemplate, ,  com.github.trojanrobotics.RobotTemplate
 MicroEdition-Profile: IMP-1.0
 MicroEdition-Configuration: CLDC-1.1
