+++ src/com/github/trojanrobotics/Config.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package edu.wpi.first.wpilibj.templates;
+package com.github.trojanrobotics;
 
 import edu.wpi.first.wpilibj.Servo;
